Created custom user details service
The email used in the JWT caused issues when updating the email because, since the email would be used to verify the token, the token would become invalid if the email were changed. This would be detrimental to the user experience. To address this, I created a new custom user details service to fully use the userID throughout the application.